Remove firewall config and allow no DNS
I wanted to use ufw and will handle that manually. I also need to handle `push "route...` things so some manual editing is required, along with routes on other machines. I also need to use the VPN as a means to restrict/protect access to other boxes, so sending DNS outside of the VPN is not a problem. This is not for a solution for all web traffic.